+++ src/metworkpy/network/centrality.py
@@ -1,6 +1,8 @@
-# The code for closeness_centrality_subset is a modified
-# version of the NetworkX closeness_centrality function
-# which is licensed under the BSD-3-Clause license
+# The code for closeness_centrality_subset, betweenness_centrality_subset,
+# _rescale, and _accumulate_subset is a modified
+# version of the NetworkX closeness_centrality, betweenness_centrality_subset,
+# _rescale, and _accumulate_subset functions.
+# NetworkX is licensed under the BSD-3-Clause license
 # Reproduced below
 # Copyright (c) 2004-2025, NetworkX Developers
 # Aric Hagberg <hagberg@lanl.gov>
@@ -45,6 +47,12 @@ import functools
 from typing import cast, Hashable, Optional, Union
 
 import networkx as nx
+from networkx.algorithms.centrality.betweenness import (
+    _single_source_dijkstra_path_basic as dijkstra,  # type: ignore
+)
+from networkx.algorithms.centrality.betweenness import (
+    _single_source_shortest_path_basic as shortest_path,  # type: ignore
+)
 
 
 def closeness_centrality_subset(
@@ -185,3 +193,153 @@ def closeness_centrality_subset(
     if u is not None:
         return closeness_dict[u]
     return closeness_dict
+
+
+def betweenness_centrality_subset(
+    G: Union[nx.Graph, nx.DiGraph],
+    targets: Optional[Iterable[Hashable]] = None,
+    normalized=True,
+    weight=None,
+):
+    r"""Compute betweenness centrality for a subset of nodes.
+
+    .. math::
+
+       c_B(v) =\sum_{s,t \in T} \frac{\sigma(s, t|v)}{\sigma(s, t)}
+
+    where $S$ is the set of sources, $T$ is the set of targets,
+    $\sigma(s, t)$ is the number of shortest $(s, t)$-paths,
+    and $\sigma(s, t|v)$ is the number of those paths
+    passing through some  node $v$ other than $s, t$.
+    If $s = t$, $\sigma(s, t) = 1$,
+    and if $v \in {s, t}$, $\sigma(s, t|v) = 0$ [2]_.
+
+    The normalization is slightly different from NetworkX,
+    as it normalizes only to the possible paths between
+    nodes in targets, not to all nodes in the network.
+
+
+    Parameters
+    ----------
+    G : graph
+      A NetworkX graph.
+
+    targets: list of nodes
+      Nodes to use as sources/targets for shortest paths in betweenness
+
+    normalized : bool, optional
+      If True the betweenness values are normalized by $2/((n-1)(n-2))$
+      for graphs, and $1/((n-1)(n-2))$ for directed graphs where $n$
+      is the number of nodes in targets.
+
+    weight : None or string, optional (default=None)
+      If None, all edge weights are considered equal.
+      Otherwise holds the name of the edge attribute used as weight.
+      Weights are used to calculate weighted shortest paths, so they are
+      interpreted as distances.
+
+    Returns
+    -------
+    nodes : dictionary
+       Dictionary of nodes with betweenness centrality as the value.
+
+    Notes
+    -----
+    The basic algorithm is from [1]_.
+
+    For weighted graphs the edge weights must be greater than zero.
+    Zero edge weights can produce an infinite number of equal length
+    paths between pairs of nodes.
+
+    The normalization might seem a little strange but it is
+    designed to make betweenness_centrality(G) be the same as
+    betweenness_centrality_subset(G,sources=G.nodes(),targets=G.nodes()).
+
+    The total number of paths between source and target is counted
+    differently for directed and undirected graphs. Directed paths
+    are easy to count. Undirected paths are tricky: should a path
+    from "u" to "v" count as 1 undirected path or as 2 directed paths?
+
+    For betweenness_centrality we report the number of undirected
+    paths when G is undirected.
+
+    For betweenness_centrality_subset the reporting is different.
+    If the source and target subsets are the same, then we want
+    to count undirected paths. But if the source and target subsets
+    differ -- for example, if sources is {0} and targets is {1},
+    then we are only counting the paths in one direction. They are
+    undirected paths but we are counting them in a directed way.
+    To count them as undirected paths, each should count as half a path.
+
+    References
+    ----------
+    .. [1] Ulrik Brandes, A Faster Algorithm for Betweenness Centrality.
+       Journal of Mathematical Sociology 25(2):163-177, 2001.
+       https://doi.org/10.1080/0022250X.2001.9990249
+    .. [2] Ulrik Brandes: On Variants of Shortest-Path Betweenness
+       Centrality and their Generic Computation.
+       Social Networks 30(2):136-145, 2008.
+       https://doi.org/10.1016/j.socnet.2007.11.001
+    """
+    b = dict.fromkeys(G, 0.0)  # b[v]=0 for v in G
+    if targets is None:
+        targets = G.nodes
+    for s in targets:
+        # single source shortest paths
+        if weight is None:  # use BFS
+            S, P, sigma, _ = shortest_path(G, s)
+        else:  # use Dijkstra's algorithm
+            S, P, sigma, _ = dijkstra(G, s, weight)
+        b = _accumulate_subset(b, S, P, sigma, s, targets)
+    b = _rescale(
+        b, set(targets), normalized=normalized, directed=G.is_directed()
+    )
+    return b
+
+
+def _rescale(
+    betweenness: dict[Hashable, float],
+    subset: set[Hashable],
+    normalized: bool,
+    directed=False,
+):
+    """
+    betweenness_centrality_subset helper.
+
+    Uses different normalization that the default in networkx
+    """
+    len_subset = len(subset)
+    if normalized:
+        if len_subset <= 2:
+            subset_scale = None  # no normalization b=0 for all nodes
+            not_subset_scale = None
+        else:
+            not_subset_scale = 1.0 / ((len_subset) * (len_subset - 1))
+            subset_scale = 1.0 / ((len_subset - 1) * (len_subset - 2))
+    else:  # rescale by 2 for undirected graphs
+        if not directed:
+            not_subset_scale = 0.5
+            subset_scale = 0.5
+        else:
+            not_subset_scale = None
+            subset_scale = None
+    if (subset_scale is not None) and (not_subset_scale is not None):
+        for v in betweenness:
+            betweenness[v] *= subset_scale if v in subset else not_subset_scale
+    return betweenness
+
+
+def _accumulate_subset(betweenness, S, P, sigma, s, targets):
+    delta = dict.fromkeys(S, 0.0)
+    target_set = set(targets) - {s}
+    while S:
+        w = S.pop()
+        if w in target_set:
+            coeff = (delta[w] + 1.0) / sigma[w]
+        else:
+            coeff = delta[w] / sigma[w]
+        for v in P[w]:
+            delta[v] += sigma[v] * coeff
+        if w != s:
+            betweenness[w] += delta[w]
+    return betweenness